Add HTML5 input field type="number".
This field provides an input type for numbers. If the browser has no HTML5 support it falls back to a simple input element like before.
You can specify, mininum value, maximum value and a step-size. You can also set it as a required field.
Be aware that you still have to validate the data in the backend!